pios: factor pios_spi completely common
