+++ flight/PiOS/Common/pios_spi.c
@@ -8,14 +8,15 @@
  *
  * @file       pios_spi.c
  * @author     The OpenPilot Team, http://www.openpilot.org Copyright (C) 2012.
- * @author     Tau Labs, http://taulabs.org, Copyright (C) 2013
+ * @author     Tau Labs, http://taulabs.org, Copyright (C) 2012-2013
+ * @author     dRonin, http://dronin.org, Copyright (C) 2016
  * @brief      Hardware Abstraction Layer for SPI ports of STM32
  * @see        The GNU Public License (GPL) Version 3
  * @notes
  *
  * Note that additional chip select lines can be easily added by using
  * the remaining free GPIOs of the core module. Shared SPI ports should be
- * arbitrated with (FreeRTOS based) Mutexes to avoid collisions!
+ * arbitrated with Mutexes to avoid collisions!
  *
  *****************************************************************************/
 /*
@@ -33,7 +34,9 @@
  * with this program; if not, write to the Free Software Foundation, Inc.,
  * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
-
+/*
+ * @todo	Clocking is wrong (interface is badly defined, should be speed not prescaler magic numbers)
+ */
 #include <pios.h>
 
 #if defined(PIOS_INCLUDE_SPI)
@@ -55,7 +58,7 @@ static struct pios_spi_dev *PIOS_SPI_alloc(void)
 * Initialises SPI pins
 * \param[in] mode currently only mode 0 supported
 * \return < 0 if initialisation failed
- */
+*/
 int32_t PIOS_SPI_Init(uint32_t *spi_id, const struct pios_spi_cfg *cfg)
 {
 	uint32_t	init_ssel = 0;
@@ -89,10 +92,12 @@ int32_t PIOS_SPI_Init(uint32_t *spi_id, const struct pios_spi_cfg *cfg)
 			SPI_NSSInternalSoftwareConfig(spi_dev->cfg->regs, SPI_NSSInternalSoft_Reset);
 		}
 		break;
+
 	case SPI_NSS_Hard:
 		/* only legal for single-slave config */
 		PIOS_Assert(spi_dev->cfg->slave_count == 1);
 		init_ssel = 1;
+		SPI_SSOutputCmd(spi_dev->cfg->regs, (spi_dev->cfg->init.SPI_Mode == SPI_Mode_Master) ? ENABLE : DISABLE);
 		/* FIXME: Should this also call SPI_SSOutputCmd()? */
 		break;
 
@@ -100,46 +105,42 @@ int32_t PIOS_SPI_Init(uint32_t *spi_id, const struct pios_spi_cfg *cfg)
 		PIOS_Assert(0);
 	}
 
+#ifndef STM32F10X_MD
 	/* Initialize the GPIO pins */
-	GPIO_Init(spi_dev->cfg->sclk.gpio, (GPIO_InitTypeDef*)&spi_dev->cfg->sclk.init);
-	GPIO_Init(spi_dev->cfg->mosi.gpio, (GPIO_InitTypeDef*)&spi_dev->cfg->mosi.init);
-	GPIO_Init(spi_dev->cfg->miso.gpio, (GPIO_InitTypeDef*)&spi_dev->cfg->miso.init);
-	for (uint32_t i = 0; i < init_ssel; i++) {
-		/* Since we're driving the SSEL pin in software, ensure that the slave is deselected */
-		/* XXX multi-slave support - maybe have another SPI_NSS_ mode? */
-		GPIO_SetBits(spi_dev->cfg->ssel[i].gpio, spi_dev->cfg->ssel[i].init.GPIO_Pin);
-		GPIO_Init(spi_dev->cfg->ssel[i].gpio, (GPIO_InitTypeDef*)&spi_dev->cfg->ssel[i].init);
+	/* note __builtin_ctz() due to the difference between GPIO_PinX and GPIO_PinSourceX */
+	if (spi_dev->cfg->remap) {
+		GPIO_PinAFConfig(spi_dev->cfg->sclk.gpio,
+		                 __builtin_ctz(spi_dev->cfg->sclk.init.GPIO_Pin),
+		                 spi_dev->cfg->remap);
+		GPIO_PinAFConfig(spi_dev->cfg->mosi.gpio,
+		                 __builtin_ctz(spi_dev->cfg->mosi.init.GPIO_Pin),
+		                 spi_dev->cfg->remap);
+		GPIO_PinAFConfig(spi_dev->cfg->miso.gpio,
+		                 __builtin_ctz(spi_dev->cfg->miso.init.GPIO_Pin),
+		                 spi_dev->cfg->remap);
+		for (uint32_t i = 0; i < init_ssel; i++) {
+			GPIO_PinAFConfig(spi_dev->cfg->ssel[i].gpio,
+			                 __builtin_ctz(spi_dev->cfg->ssel[i].init.GPIO_Pin),
+			                 spi_dev->cfg->remap);
+		}
 	}
+#endif
+	GPIO_Init(spi_dev->cfg->sclk.gpio, (GPIO_InitTypeDef *) & (spi_dev->cfg->sclk.init));
+	GPIO_Init(spi_dev->cfg->mosi.gpio, (GPIO_InitTypeDef *) & (spi_dev->cfg->mosi.init));
+	GPIO_Init(spi_dev->cfg->miso.gpio, (GPIO_InitTypeDef *) & (spi_dev->cfg->miso.init));
 
-	/* Enable the associated peripheral clock */
-	switch ((uint32_t) spi_dev->cfg->regs) {
-	case (uint32_t) SPI1:
-		/* Enable SPI peripheral clock (APB2 == high speed) */
-		RCC_APB2PeriphClockCmd(RCC_APB2Periph_SPI1, ENABLE);
-		break;
-	case (uint32_t) SPI2:
-		/* Enable SPI peripheral clock (APB1 == slow speed) */
-		RCC_APB1PeriphClockCmd(RCC_APB1Periph_SPI2, ENABLE);
-		break;
-	case (uint32_t) SPI3:
-		/* Enable SPI peripheral clock (APB1 == slow speed) */
-		RCC_APB1PeriphClockCmd(RCC_APB1Periph_SPI3, ENABLE);
-		break;
+	if (spi_dev->cfg->init.SPI_NSS != SPI_NSS_Hard) {
+		for (uint32_t i = 0; i < init_ssel; i++) {
+			/* Since we're driving the SSEL pin in software, ensure that the slave is deselected */
+			/* XXX multi-slave support - maybe have another SPI_NSS_ mode? */
+			GPIO_SetBits(spi_dev->cfg->ssel[i].gpio, spi_dev->cfg->ssel[i].init.GPIO_Pin);
+			GPIO_Init(spi_dev->cfg->ssel[i].gpio, (GPIO_InitTypeDef *) & (spi_dev->cfg->ssel[i].init));
+		}
 	}
-
-	/* Enable DMA clock */
-	RCC_AHBPeriphClockCmd(spi_dev->cfg->dma.ahb_clk, ENABLE);
-
-	/* Configure DMA for SPI Rx */
-	DMA_Cmd(spi_dev->cfg->dma.rx.channel, DISABLE);
-	DMA_Init(spi_dev->cfg->dma.rx.channel, (DMA_InitTypeDef*)&spi_dev->cfg->dma.rx.init);
-
-	/* Configure DMA for SPI Tx */
-	DMA_Cmd(spi_dev->cfg->dma.tx.channel, DISABLE);
-	DMA_Init(spi_dev->cfg->dma.tx.channel, (DMA_InitTypeDef*)&spi_dev->cfg->dma.tx.init);
 
 	/* Initialize the SPI block */
-	SPI_Init(spi_dev->cfg->regs, (SPI_InitTypeDef*)&spi_dev->cfg->init);
+	SPI_I2S_DeInit(spi_dev->cfg->regs);
+	SPI_Init(spi_dev->cfg->regs, (SPI_InitTypeDef *) & (spi_dev->cfg->init));
 
 	/* Configure CRC calculation */
 	if (spi_dev->cfg->use_crc) {
@@ -151,13 +152,9 @@ int32_t PIOS_SPI_Init(uint32_t *spi_id, const struct pios_spi_cfg *cfg)
 	/* Enable SPI */
 	SPI_Cmd(spi_dev->cfg->regs, ENABLE);
 
-	/* Enable SPI interrupts to DMA */
-	SPI_I2S_DMACmd(spi_dev->cfg->regs, SPI_I2S_DMAReq_Tx | SPI_I2S_DMAReq_Rx, ENABLE);
-
-	/* Configure DMA interrupt */
-	NVIC_Init((NVIC_InitTypeDef*)&spi_dev->cfg->dma.irq.init);
-
+	/* Must store this before enabling interrupt */
 	*spi_id = (uint32_t)spi_dev;
+
 	return (0);
 
 out_fail:
@@ -165,37 +162,45 @@ out_fail:
 }
 
 /**
-* (Re-)initialises SPI peripheral clock rate
-*
-* \param[in] spi SPI number (0 or 1)
-* \param[in] spi_prescaler configures the SPI speed:
-* <UL>
-*   <LI>PIOS_SPI_PRESCALER_2: sets clock rate 27.7~ nS @ 72 MHz (36 MBit/s) (only supported for spi==0, spi1 uses 4 instead)
-*   <LI>PIOS_SPI_PRESCALER_4: sets clock rate 55.5~ nS @ 72 MHz (18 MBit/s)
-*   <LI>PIOS_SPI_PRESCALER_8: sets clock rate 111.1~ nS @ 72 MHz (9 MBit/s)
-*   <LI>PIOS_SPI_PRESCALER_16: sets clock rate 222.2~ nS @ 72 MHz (4.5 MBit/s)
-*   <LI>PIOS_SPI_PRESCALER_32: sets clock rate 444.4~ nS @ 72 MHz (2.25 MBit/s)
-*   <LI>PIOS_SPI_PRESCALER_64: sets clock rate 888.8~ nS @ 72 MHz (1.125 MBit/s)
-*   <LI>PIOS_SPI_PRESCALER_128: sets clock rate 1.7~ nS @ 72 MHz (0.562 MBit/s)
-*   <LI>PIOS_SPI_PRESCALER_256: sets clock rate 3.5~ nS @ 72 MHz (0.281 MBit/s)
-* </UL>
-* \return 0 if no error
-* \return -1 if disabled SPI port selected
-* \return -3 if invalid spi_prescaler selected
-*/
+ * (Re-)initialises SPI peripheral clock rate
+ *
+ * \param[in] spi SPI number (0 or 1)
+ * \param[in] spi_prescaler configures the SPI speed:
+ * <UL>
+ *   <LI>PIOS_SPI_PRESCALER_2: sets clock rate 27.7~ nS @ 72 MHz (36 MBit/s) (only supported for spi==0, spi1 uses 4 instead)
+ *   <LI>PIOS_SPI_PRESCALER_4: sets clock rate 55.5~ nS @ 72 MHz (18 MBit/s)
+ *   <LI>PIOS_SPI_PRESCALER_8: sets clock rate 111.1~ nS @ 72 MHz (9 MBit/s)
+ *   <LI>PIOS_SPI_PRESCALER_16: sets clock rate 222.2~ nS @ 72 MHz (4.5 MBit/s)
+ *   <LI>PIOS_SPI_PRESCALER_32: sets clock rate 444.4~ nS @ 72 MHz (2.25 MBit/s)
+ *   <LI>PIOS_SPI_PRESCALER_64: sets clock rate 888.8~ nS @ 72 MHz (1.125 MBit/s)
+ *   <LI>PIOS_SPI_PRESCALER_128: sets clock rate 1.7~ nS @ 72 MHz (0.562 MBit/s)
+ *   <LI>PIOS_SPI_PRESCALER_256: sets clock rate 3.5~ nS @ 72 MHz (0.281 MBit/s)
+ * </UL>
+ * \return 0 if no error
+ * \return -1 if disabled SPI port selected
+ * \return -3 if invalid spi_prescaler selected
+ */
 int32_t PIOS_SPI_SetClockSpeed(uint32_t spi_id, uint32_t spi_speed)
 {
 	struct pios_spi_dev *spi_dev = (struct pios_spi_dev *)spi_id;
 
 	bool valid = PIOS_SPI_validate(spi_dev);
-	PIOS_Assert(valid)
+	PIOS_Assert(valid);
 
-	SPI_InitTypeDef	SPI_InitStructure;
+	SPI_InitTypeDef SPI_InitStructure;
 
 	SPIPrescalerTypeDef spi_prescaler;
 
+	//SPI clock is different depending on the bus
+	uint32_t spiBusClock = 0;
+
+	if(spi_dev->cfg->regs == SPI1)
+		spiBusClock = PIOS_SYSCLK / 2;
+	else
+		spiBusClock = PIOS_SYSCLK;
+
 	//The needed prescaler for desired speed
-	float desiredPrescaler = (float) PIOS_SYSCLK / spi_speed;
+	float desiredPrescaler=(float) spiBusClock / spi_speed;
 
 	//Choosing the existing prescaler nearest the desiredPrescaler
 	if(desiredPrescaler <= 2)
@@ -227,7 +232,7 @@ int32_t PIOS_SPI_SetClockSpeed(uint32_t spi_id, uint32_t spi_speed)
 	PIOS_SPI_TransferByte(spi_id, 0xFF);
 
 	//return set speed
-	return PIOS_SYSCLK / spi_prescaler;
+	return spiBusClock / spi_prescaler;
 }
 
 /**
@@ -269,6 +274,7 @@ int32_t PIOS_SPI_ClaimBusISR(uint32_t spi_id, bool *woken)
 	return 0;
 }
 
+
 /**
  * Release the SPI bus semaphore.  Calling the SPI functions does not require this
  * \param[in] spi SPI number (0 or 1)
@@ -305,11 +311,11 @@ int32_t PIOS_SPI_ReleaseBusISR(uint32_t spi_id, bool *woken)
 }
 
 /**
- * Controls the RC (Register Clock alias Chip Select) pin of a SPI port
- * \param[in] spi SPI number (0 or 1)
- * \param[in] pin_value 0 or 1
- * \return 0 if no error
- */
+* Controls the RC (Register Clock alias Chip Select) pin of a SPI port
+* \param[in] spi SPI number (0 or 1)
+* \param[in] pin_value 0 or 1
+* \return 0 if no error
+*/
 int32_t PIOS_SPI_RC_PinSet(uint32_t spi_id, uint32_t slave_id, uint8_t pin_value)
 {
 	struct pios_spi_dev *spi_dev = (struct pios_spi_dev *)spi_id;
@@ -333,7 +339,6 @@ int32_t PIOS_SPI_RC_PinSet(uint32_t spi_id, uint32_t slave_id, uint8_t pin_value
 * \param[in] spi SPI number (0 or 1)
 * \param[in] b the byte which should be transfered
 */
-static uint8_t dummy;
 int32_t PIOS_SPI_TransferByte(uint32_t spi_id, uint8_t b)
 {
 	struct pios_spi_dev *spi_dev = (struct pios_spi_dev *)spi_id;
@@ -348,30 +353,34 @@ int32_t PIOS_SPI_TransferByte(uint32_t spi_id, uint8_t b)
 	 */
 
 	/* Make sure the RXNE flag is cleared by reading the DR register */
-	dummy = spi_dev->cfg->regs->DR;
+	(void)spi_dev->cfg->regs->DR;
+
+	/* Wait until the TXE goes high */
+	while (SPI_I2S_GetFlagStatus(spi_dev->cfg->regs, SPI_I2S_FLAG_TXE) == RESET);
 
 	/* Start the transfer */
 	spi_dev->cfg->regs->DR = b;
 
 	/* Wait until there is a byte to read */
-	while (!(spi_dev->cfg->regs->SR & SPI_I2S_FLAG_RXNE)) ;
+	while (SPI_I2S_GetFlagStatus(spi_dev->cfg->regs, SPI_I2S_FLAG_RXNE) == RESET);
 
 	/* Read the rx'd byte */
 	rx_byte = spi_dev->cfg->regs->DR;
 
 	/* Wait until the TXE goes high */
-	while (!(spi_dev->cfg->regs->SR & SPI_I2S_FLAG_TXE)) ;
+	while (!(spi_dev->cfg->regs->SR & SPI_I2S_FLAG_TXE));
 
 	/* Wait for SPI transfer to have fully completed */
-	while (spi_dev->cfg->regs->SR & SPI_I2S_FLAG_BSY) ;
+	while (spi_dev->cfg->regs->SR & SPI_I2S_FLAG_BSY);
 
 	/* Return received byte */
 	return rx_byte;
 }
 
 /**
-* Transfers a block of bytes via DMA.  Blocks until transfer completed.
-* \param[in] spi SPI number (0 or 1)
+* Transfers a block of bytes via PIO.
+*
+* \param[in] spi_id SPI device handle
 * \param[in] send_buffer pointer to buffer which should be sent.<BR>
 * If NULL, 0xff (all-one) will be sent.
 * \param[in] receive_buffer pointer to buffer which should get the received values.<BR>
@@ -379,129 +388,64 @@ int32_t PIOS_SPI_TransferByte(uint32_t spi_id, uint8_t b)
 * \param[in] len number of bytes which should be transfered
 * \return >= 0 if no error during transfer
 * \return -1 if disabled SPI port selected
-* \return -3 if function has been called during an ongoing DMA transfer
 */
-int32_t PIOS_SPI_TransferBlock(uint32_t spi_id, const uint8_t *send_buffer, uint8_t *receive_buffer, uint16_t len)
+static int32_t SPI_PIO_TransferBlock(uint32_t spi_id, const uint8_t *send_buffer, uint8_t *receive_buffer, uint16_t len)
 {
 	struct pios_spi_dev *spi_dev = (struct pios_spi_dev *)spi_id;
+	uint8_t b;
 
 	bool valid = PIOS_SPI_validate(spi_dev);
 	PIOS_Assert(valid)
 
-	DMA_InitTypeDef       dma_init;
+	while (len--) {
+		/* get the byte to send */
+		b = send_buffer ? *(send_buffer++) : 0xff;
 
-	/* Exit if ongoing transfer */
-	if (DMA_GetCurrDataCounter(spi_dev->cfg->dma.rx.channel)) {
-		return -3;
+		/* Wait until the TXE goes high */
+		while (SPI_I2S_GetFlagStatus(spi_dev->cfg->regs, SPI_I2S_FLAG_TXE) == RESET);
+
+		/* Start the transfer */
+		spi_dev->cfg->regs->DR = b;
+
+		/* Wait until there is a byte to read */
+		while (SPI_I2S_GetFlagStatus(spi_dev->cfg->regs, SPI_I2S_FLAG_RXNE) == RESET);
+
+		/* Read the rx'd byte */
+		b = spi_dev->cfg->regs->DR;
+
+		/* save the received byte */
+		if (receive_buffer)
+			*(receive_buffer++) = b;
 	}
 
-	/* Disable the SPI peripheral */
-	SPI_Cmd(spi_dev->cfg->regs, DISABLE);
+	/* Wait for SPI transfer to have fully completed */
+	while (spi_dev->cfg->regs->SR & SPI_I2S_FLAG_BSY);
 
-	/* Disable the DMA channels */
-	DMA_Cmd(spi_dev->cfg->dma.rx.channel, DISABLE);
-	DMA_Cmd(spi_dev->cfg->dma.tx.channel, DISABLE);
-
-	/*
-	 * Configure Rx channel
-	 */
-
-	/* Start with the default configuration for this peripheral */
-	dma_init = spi_dev->cfg->dma.rx.init;
-
-	if (receive_buffer != NULL) {
-		/* Enable memory addr. increment - bytes written into receive buffer */
-		dma_init.DMA_MemoryBaseAddr = (uint32_t) receive_buffer;
-		dma_init.DMA_MemoryInc = DMA_MemoryInc_Enable;
-	} else {
-		/* Disable memory addr. increment - bytes written into dummy buffer */
-		spi_dev->rx_dummy_byte = 0xFF;
-		dma_init.DMA_MemoryBaseAddr = (uint32_t) &spi_dev->rx_dummy_byte;
-		dma_init.DMA_MemoryInc = DMA_MemoryInc_Disable;
-	}
-	if (spi_dev->cfg->use_crc) {
-		/* Make sure the CRC error flag is cleared before we start */
-		SPI_I2S_ClearFlag(spi_dev->cfg->regs, SPI_FLAG_CRCERR);
-	}
-
-	dma_init.DMA_BufferSize = len;
-	DMA_Init(spi_dev->cfg->dma.rx.channel, &(dma_init));
-
-	/*
-	 * Configure Tx channel
-	 */
-
-	/* Start with the default configuration for this peripheral */
-	dma_init = spi_dev->cfg->dma.tx.init;
-
-	if (send_buffer != NULL) {
-		/* Enable memory addr. increment - bytes written into receive buffer */
-		dma_init.DMA_MemoryBaseAddr = (uint32_t) send_buffer;
-		dma_init.DMA_MemoryInc = DMA_MemoryInc_Enable;
-	} else {
-		/* Disable memory addr. increment - bytes written into dummy buffer */
-		spi_dev->tx_dummy_byte = 0xFF;
-		dma_init.DMA_MemoryBaseAddr = (uint32_t) &spi_dev->tx_dummy_byte;
-		dma_init.DMA_MemoryInc = DMA_MemoryInc_Disable;
-	}
-
-	if (spi_dev->cfg->use_crc) {
-		/* The last byte of the payload will be replaced with the CRC8 */
-		dma_init.DMA_BufferSize = len - 1;
-	} else {
-		dma_init.DMA_BufferSize = len;
-	}
-
-	DMA_Init(spi_dev->cfg->dma.tx.channel, &(dma_init));
-
-	/* Flush out the CRC registers */
-	SPI_CalculateCRC(spi_dev->cfg->regs, DISABLE);
-	(void)SPI_GetCRC(spi_dev->cfg->regs, SPI_CRC_Rx);
-	SPI_I2S_ClearFlag(spi_dev->cfg->regs, SPI_FLAG_CRCERR);
-
-	/* Make sure to flush out the receive buffer */
-	(void)SPI_I2S_ReceiveData(spi_dev->cfg->regs);
-
-	if (spi_dev->cfg->use_crc) {
-		/* Need a 0->1 transition to reset the CRC logic */
-		SPI_CalculateCRC(spi_dev->cfg->regs, ENABLE);
-	}
-
-	/* Start DMA transfers */
-	DMA_Cmd(spi_dev->cfg->dma.rx.channel, ENABLE);
-	DMA_Cmd(spi_dev->cfg->dma.tx.channel, ENABLE);
-
-	/* Reenable the SPI device */
-	SPI_Cmd(spi_dev->cfg->regs, ENABLE);
-
-	/* Wait until all bytes have been transmitted/received */
-	while (DMA_GetCurrDataCounter(spi_dev->cfg->dma.rx.channel));
-
-	/* Wait for the final bytes of the transfer to complete, including CRC byte(s). */
-	while (!(SPI_I2S_GetFlagStatus(spi_dev->cfg->regs, SPI_I2S_FLAG_TXE)));
-
-	/* Wait for the final bytes of the transfer to complete, including CRC byte(s). */
-	while (SPI_I2S_GetFlagStatus(spi_dev->cfg->regs, SPI_I2S_FLAG_BSY));
-
-	/* Check the CRC on the transfer if enabled. */
-	if (spi_dev->cfg->use_crc) {
-		/* Check the SPI CRC error flag */
-		if (SPI_I2S_GetFlagStatus(spi_dev->cfg->regs, SPI_FLAG_CRCERR)) {
-			return -4;
-		}
-	}
-
-	/* No error */
 	return 0;
+}
+
+
+/**
+* Transfers a block of bytes
+* \param[in] spi_id SPI device handle
+* \param[in] send_buffer pointer to buffer which should be sent.<BR>
+* If NULL, 0xff (all-one) will be sent.
+* \param[in] receive_buffer pointer to buffer which should get the received values.<BR>
+* If NULL, received bytes will be discarded.
+* \param[in] len number of bytes which should be transfered
+* \return >= 0 if no error during transfer
+* \return -1 if disabled SPI port selected
+*/
+int32_t PIOS_SPI_TransferBlock(uint32_t spi_id, const uint8_t *send_buffer, uint8_t *receive_buffer, uint16_t len)
+{
+	return SPI_PIO_TransferBlock(spi_id, send_buffer, receive_buffer, len);
 }
 
 /**
 * Check if a transfer is in progress
 * \param[in] spi SPI number (0 or 1)
 * \return >= 0 if no transfer is in progress
-* \return -1 if disabled SPI port selected
-* \return -2 if unsupported SPI port selected
-* \return -3 if function has been called during an ongoing DMA transfer
+* \return -3 if transfer in progress
 */
 int32_t PIOS_SPI_Busy(uint32_t spi_id)
 {
@@ -510,9 +454,7 @@ int32_t PIOS_SPI_Busy(uint32_t spi_id)
 	bool valid = PIOS_SPI_validate(spi_dev);
 	PIOS_Assert(valid)
 
-	/* DMA buffer has data or SPI transmit register not empty or SPI is busy*/
-	if (DMA_GetCurrDataCounter(spi_dev->cfg->dma.rx.channel) ||
-	    !SPI_I2S_GetFlagStatus(spi_dev->cfg->regs, SPI_I2S_FLAG_TXE) ||
+	if (!SPI_I2S_GetFlagStatus(spi_dev->cfg->regs, SPI_I2S_FLAG_TXE) ||
 	    SPI_I2S_GetFlagStatus(spi_dev->cfg->regs, SPI_I2S_FLAG_BSY)) {
 		return -3;
 	}
@@ -520,31 +462,28 @@ int32_t PIOS_SPI_Busy(uint32_t spi_id)
 	return (0);
 }
 
-void PIOS_SPI_SetPrescalar(uint32_t spi_id, uint32_t prescaler)
-{
-	struct pios_spi_dev *spi_dev = (struct pios_spi_dev *)spi_id;
-
-	bool valid = PIOS_SPI_validate(spi_dev);
-	PIOS_Assert(valid);
-	PIOS_Assert(IS_SPI_BAUDRATE_PRESCALER(prescaler));
-
-	spi_dev->cfg->regs->CR1 = (spi_dev->cfg->regs->CR1 & ~0x0038) | prescaler;
-}
-
 void PIOS_SPI_IRQ_Handler(uint32_t spi_id)
 {
+#if defined(PIOS_INCLUDE_CHIBIOS)
+	CH_IRQ_PROLOGUE();
+#endif /* defined(PIOS_INCLUDE_CHIBIOS) */
+
 	struct pios_spi_dev *spi_dev = (struct pios_spi_dev *)spi_id;
 
 	bool valid = PIOS_SPI_validate(spi_dev);
 	PIOS_Assert(valid)
 
-	DMA_ClearFlag(spi_dev->cfg->dma.irq.flags);
+	if (spi_dev->cfg->init.SPI_Mode == SPI_Mode_Master) {
+		/* Wait for the final bytes of the transfer to complete, including CRC byte(s). */
+		while (!(SPI_I2S_GetFlagStatus(spi_dev->cfg->regs, SPI_I2S_FLAG_TXE))) ;
 
-	/* Wait for the final bytes of the transfer to complete, including CRC byte(s). */
-	while (!(SPI_I2S_GetFlagStatus(spi_dev->cfg->regs, SPI_I2S_FLAG_TXE))) ;
+		/* Wait for the final bytes of the transfer to complete, including CRC byte(s). */
+		while (SPI_I2S_GetFlagStatus(spi_dev->cfg->regs, SPI_I2S_FLAG_BSY)) ;
+	}
 
-	/* Wait for the final bytes of the transfer to complete, including CRC byte(s). */
-	while (SPI_I2S_GetFlagStatus(spi_dev->cfg->regs, SPI_I2S_FLAG_BSY)) ;
+#if defined(PIOS_INCLUDE_CHIBIOS)
+	CH_IRQ_EPILOGUE();
+#endif /* defined(PIOS_INCLUDE_CHIBIOS) */
 }
 
 #endif
