+++ flight/targets/discoveryf4/fw/Makefile
@@ -115,6 +115,7 @@ SRC += $(PIOSCOMMON)/pios_queue.c
 SRC += $(PIOSCOMMON)/pios_hal.c
 SRC += $(PIOSCOMMON)/pios_servo.c
 SRC += $(PIOSCOMMON)/pios_modules.c
+SRC += $(PIOSCOMMON)/pios_spi.c
 
 # List any extra directories to look for include files here.
 #    Each directory must be seperated by a space.
